requirements.txt
numpy
matplotlib
dlib
opencv-python
torch
torchvision
sphinx
sphinx-rtd-theme
scikit-learn
scikit-image
Pillow